# Compress optional operators into separate rules