… Sources/SynKit/Tests/Parser.Generator.Tests/verified/ExpressionParserTests.SimpleMathExpression.verified.txt