since that faster then trying each alternative one by one.

## Before

| Method                 | Mean           | Error        | StdDev       |
|----------------------- |---------------:|-------------:|-------------:|
| ExpressionParser       | 3,369,936.9 us | 35,774.27 us | 29,873.13 us |
| ManualExpressionParser |    13,656.3 us |    196.00 us |    183.34 us |
| Lex                    |       955.6 us |      8.25 us |      7.72 us |

## After 

| Method                 | Mean           | Error         | StdDev        | Median         |
|----------------------- |---------------:|--------------:|--------------:|---------------:|
| ExpressionParser       | 3,191,005.3 us | 104,642.82 us | 301,918.49 us | 3,024,123.9 us |
| ManualExpressionParser |    13,627.5 us |     166.75 us |     147.82 us |    13,611.9 us |
| Lex                    |       975.2 us |      15.86 us |      17.63 us |       973.1 us |